Remove pom_generator to clean up, and add aruba tests.
File generation from a template has been generalized to be used
in places other than pom file, and is now made available to tasks
by merely including the `TaskHelper` module.
The aruba integration tests will ensure that the templates are
properly defined so that the output matches the expectations.  This
particular task yet has to be done.  Until it is doneI am leaving
`generator_spec` in the codebase, with its content commented out,
as it is going to be the base for these integration tests.